--- content.json
@@ -1,125 +1,0 @@
-{
-	"tabs": [
-		{
-			"name": "debugTab",
-			"label": "Debug",
-			"buttons": [
-				{ "id": "hsCacheBuster", "label": "Bust Cache", "width": "50" },
-				{ "id": "hsDebug", "label": "Debug Mode", "width": "50" },
-				{ "id": "hsMoveJQueryToFooter", "label": "Footer jQuery", "width": "50" },
-				{ "id": "hsAmp", "label": "AMP", "width": "50" },
-				{
-					"id": "pageSpeed",
-					"label": "Google Page Speed",
-					"width": "100"
-				},
-				{
-					"id": "hsOpenDesignManager",
-					"label": "Open Design Manager",
-					"width": "100",
-					"url": "https://app.hubspot.com/l/design-manager"
-				}
-			]
-		},
-		{
-			"name": "developerTab",
-			"label": "Developer",
-			"buttons": [{ "id": "developerMode", "label": "Developer Mode", "width": "100" }],
-			"toggles": [
-				{ "id": "uitweaks", "label": "Dev Menu" },
-				{ "id": "darktheme", "label": "Dark Theme" }
-			],
-			"links": [
-				{
-					"label": "Documentation",
-					"children": [
-						{
-							"label": "Functions",
-							"url": "https://developers.hubspot.com/docs/cms/hubl/functions"
-						},
-						{
-							"label": "Variables",
-							"url": "https://developers.hubspot.com/docs/cms/hubl/variables"
-						},
-						{
-							"label": "Filters",
-							"url": "https://developers.hubspot.com/docs/cms/hubl/filters"
-						},
-						{
-							"label": "HubDB",
-							"url": "https://developers.hubspot.com/docs/cms/data/hubdb"
-						}
-					]
-				},
-				{
-					"label": "Community",
-					"children": [
-						{
-							"label": "Slack",
-							"url": "https://developers.hubspot.com/docs/cms/hubl/functions"
-						},
-						{
-							"label": "Forum",
-							"url": "https://community.hubspot.com/"
-						},
-						{
-							"label": "Ideas",
-							"url": "https://community.hubspot.com/t5/HubSpot-Ideas/idb-p/HubSpot_Ideas"
-						}
-					]
-				}
-			]
-		}
-	],
-	"devMenu": [
-		{
-			"label": "Design Manager",
-			"url": "design-manager/_HUB_ID_"
-		},
-		{
-			"label": "Content Staging",
-			"url": "content/_HUB_ID_/staging/"
-		},
-		{
-			"label": "Website Pages",
-			"url": "website/_HUB_ID_/pages/site/all"
-		},
-		{
-			"label": "Landing Pages",
-			"url": "website/_HUB_ID_/pages/landing/all"
-		},
-		{
-			"label": "Blogs",
-			"url": "website/_HUB_ID_/blog/posts"
-		},
-		{
-			"label": "HubDB",
-			"url": "hubdb/_HUB_ID_"
-		},
-		{
-			"label": "File Manager",
-			"url": "files/_HUB_ID_"
-		},
-		{
-			"label": "Advanced Menus",
-			"url": "settings/_HUB_ID_/website/navigation"
-		},
-		{
-			"label": "Content Settings",
-			"url": "settings/_HUB_ID_/website/pages/all-domains/page-templates"
-		},
-		{
-			"label": "URL Redirects",
-			"url": "settings/_HUB_ID_/domains/url-redirects"
-		},
-		{
-			"label": "Marketplace",
-			"url": "ecosystem/_HUB_ID_/marketplace/templates?source=topNav"
-		},
-		{
-			"label": "GraphiQL",
-			"url": "graphiql/_HUB_ID_"
-		}
-	],
-	"options": []
-}
