--- popup/popup.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-	<head>
-		<title>HS Dev</title>
-		<link rel="stylesheet" type="text/css" href="popup.css" />
-	</head>
-	<body>
-		<button id="open-options-button">Open Options</button>
-		<div class="tabs" id="tabsContainer"></div>
-		<div id="tabContentContainer"></div>
-
-		<script src="popup.js"></script>
-	</body>
-</html>
