+++ src/content.json
@@ -1,0 +1,53 @@
+{
+	"devMenu": [
+		{
+			"label": "Design Manager",
+			"url": "design-manager/_HUB_ID_"
+		},
+		{
+			"label": "Content Staging",
+			"url": "content/_HUB_ID_/staging/"
+		},
+		{
+			"label": "Website Pages",
+			"url": "website/_HUB_ID_/pages/site/all"
+		},
+		{
+			"label": "Landing Pages",
+			"url": "website/_HUB_ID_/pages/landing/all"
+		},
+		{
+			"label": "Blogs",
+			"url": "website/_HUB_ID_/blog/posts"
+		},
+		{
+			"label": "HubDB",
+			"url": "hubdb/_HUB_ID_"
+		},
+		{
+			"label": "File Manager",
+			"url": "files/_HUB_ID_"
+		},
+		{
+			"label": "Advanced Menus",
+			"url": "settings/_HUB_ID_/website/navigation"
+		},
+		{
+			"label": "Content Settings",
+			"url": "settings/_HUB_ID_/website/pages/all-domains/page-templates"
+		},
+		{
+			"label": "URL Redirects",
+			"url": "settings/_HUB_ID_/domains/url-redirects"
+		},
+		{
+			"label": "Marketplace",
+			"url": "ecosystem/_HUB_ID_/marketplace/templates?source=topNav"
+		},
+		{
+			"label": "GraphiQL",
+			"url": "graphiql/_HUB_ID_"
+		}
+	],
+	"options": []
+}
