+++ src/css/input.css
@@ -1,0 +1,27 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+@layer base {
+    body {
+        width: 600px;
+        @apply p-10
+    }
+}
+@layer components {
+    .btn {
+        @apply rounded bg-white px-2 py-1 text-xs font-semibold text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 hover:bg-gray-50
+    }
+    .tab-resting {
+        @apply text-gray-500 hover:text-gray-700 relative min-w-0 flex-1 overflow-hidden bg-white py-4 px-4 text-center text-sm font-medium hover:bg-gray-50 focus:z-10 cursor-pointer
+    }
+    .tab-active {
+        @apply text-gray-900 rounded-l-lg relative min-w-0 flex-1 overflow-hidden bg-white py-4 px-4 text-center text-sm font-medium hover:bg-gray-50 focus:z-10
+    }
+
+    .tab-resting span.absolute {
+        @apply bg-transparent
+    }
+    .tab-active span.absolute {
+        @apply bg-indigo-500
+    }
+}
