+++ src/devtools/panel.html
