+++ src/popup/hsDarkIde.css
