+++ src/popup/popup.html
@@ -1,0 +1,167 @@
+<!DOCTYPE html>
+<html>
+	<head>
+		<title>HS Dev</title>
+		<link href="../css/output.css" rel="stylesheet" />
+	</head>
+	<body>
+		<button id="open-options-button" class="text-gray-900" type="button">
+			<svg
+				fill="currentColor"
+				xmlns="http://www.w3.org/2000/svg"
+				class="h-5 w-5"
+				viewBox="0 0 512 512"
+			>
+				<!--! Font Awesome Free 6.4.2 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license (Commercial License) Copyright 2023 Fonticons, Inc. -->
+				<path
+					d="M495.9 166.6c3.2 8.7 .5 18.4-6.4 24.6l-43.3 39.4c1.1 8.3 1.7 16.8 1.7 25.4s-.6 17.1-1.7 25.4l43.3 39.4c6.9 6.2 9.6 15.9 6.4 24.6c-4.4 11.9-9.7 23.3-15.8 34.3l-4.7 8.1c-6.6 11-14 21.4-22.1 31.2c-5.9 7.2-15.7 9.6-24.5 6.8l-55.7-17.7c-13.4 10.3-28.2 18.9-44 25.4l-12.5 57.1c-2 9.1-9 16.3-18.2 17.8c-13.8 2.3-28 3.5-42.5 3.5s-28.7-1.2-42.5-3.5c-9.2-1.5-16.2-8.7-18.2-17.8l-12.5-57.1c-15.8-6.5-30.6-15.1-44-25.4L83.1 425.9c-8.8 2.8-18.6 .3-24.5-6.8c-8.1-9.8-15.5-20.2-22.1-31.2l-4.7-8.1c-6.1-11-11.4-22.4-15.8-34.3c-3.2-8.7-.5-18.4 6.4-24.6l43.3-39.4C64.6 273.1 64 264.6 64 256s.6-17.1 1.7-25.4L22.4 191.2c-6.9-6.2-9.6-15.9-6.4-24.6c4.4-11.9 9.7-23.3 15.8-34.3l4.7-8.1c6.6-11 14-21.4 22.1-31.2c5.9-7.2 15.7-9.6 24.5-6.8l55.7 17.7c13.4-10.3 28.2-18.9 44-25.4l12.5-57.1c2-9.1 9-16.3 18.2-17.8C227.3 1.2 241.5 0 256 0s28.7 1.2 42.5 3.5c9.2 1.5 16.2 8.7 18.2 17.8l12.5 57.1c15.8 6.5 30.6 15.1 44 25.4l55.7-17.7c8.8-2.8 18.6-.3 24.5 6.8c8.1 9.8 15.5 20.2 22.1 31.2l4.7 8.1c6.1 11 11.4 22.4 15.8 34.3zM256 336a80 80 0 1 0 0-160 80 80 0 1 0 0 160z"
+				/>
+			</svg>
+		</button>
+		<div>
+			<nav class="isolate flex divide-x divide-gray-200 rounded-lg shadow" aria-label="Tabs">
+				<a class="tab-resting tab-active" data-tab="tab1">
+					<span>Debugging</span>
+					<span aria-hidden="true" class="absolute inset-x-0 bottom-0 h-0.5"></span>
+				</a>
+				<a class="tab-resting" data-tab="tab2">
+					<span>Devloper</span>
+					<span aria-hidden="true" class="absolute inset-x-0 bottom-0 h-0.5"></span>
+				</a>
+				<a class="tab-resting" data-tab="tab3">
+					<span>Links</span>
+					<span
+						aria-hidden="true"
+						class="bg-transparent absolute inset-x-0 bottom-0 h-0.5"
+					></span>
+				</a>
+			</nav>
+
+			<div id="tab1" class="tab-content">
+				<button class="btn hsButton" id="hsCacheBuster">Bust Cache</button>
+				<button class="btn hsButton" id="hsDebug">Debug Mode</button>
+				<button class="btn hsButton" id="hsMoveJQueryToFooter">Footer jQuery</button>
+				<button class="btn hsButton" id="hsAmp">AMP</button>
+				<button class="btn hsButton" id="pageSpeed">Google Page Speed</button>
+				<button
+					class="btn hsButton"
+					id="hsOpenDesignManager"
+					data-url="https://app.hubspot.com/l/design-manager"
+				>
+					Open Design Manager
+				</button>
+			</div>
+			<div id="tab2" class="tab-content hidden">
+				<button class="btn hsButton" id="developerMode">Developer Mode</button>
+
+				<div class="flex items-center justify-between">
+					<span class="flex flex-grow flex-col">
+						<span
+							class="text-sm font-medium leading-6 text-gray-900"
+							id="devMenu-label"
+						>
+							Developer Menu
+						</span>
+						<span class="text-sm text-gray-500" id="devMenu-description">
+							Enable developer menu in the HubSpot menu bar.
+						</span>
+					</span>
+					<!-- Enabled: "bg-indigo-600", Not Enabled: "bg-gray-200" -->
+					<button
+						data-storage="devMenu"
+						type="button"
+						class="hsToggle bg-gray-200 relative inline-flex h-6 w-11 flex-shrink-0 cursor-pointer rounded-full border-2 border-transparent transition-colors duration-200 ease-in-out focus:outline-none focus:ring-2 focus:ring-indigo-600 focus:ring-offset-2"
+						role="switch"
+						aria-checked="false"
+						aria-labelledby="devMenu-label"
+						aria-describedby="devMenu-description"
+					>
+						<!-- Enabled: "translate-x-5", Not Enabled: "translate-x-0" -->
+						<span
+							aria-hidden="true"
+							class="translate-x-0 pointer-events-none inline-block h-5 w-5 transform rounded-full bg-white shadow ring-0 transition duration-200 ease-in-out"
+						></span>
+					</button>
+				</div>
+
+				<div class="flex items-center justify-between">
+					<span class="flex flex-grow flex-col">
+						<span
+							class="text-sm font-medium leading-6 text-gray-900"
+							id="darktheme-label"
+						>
+							IDE Dark Theme
+						</span>
+						<span class="text-sm text-gray-500" id="darktheme-description">
+							Enable dark theme in the HubSpot IDE.
+						</span>
+					</span>
+					<!-- Enabled: "bg-indigo-600", Not Enabled: "bg-gray-200" -->
+					<button
+						data-storage="darktheme"
+						type="button"
+						class="hsToggle bg-gray-200 relative inline-flex h-6 w-11 flex-shrink-0 cursor-pointer rounded-full border-2 border-transparent transition-colors duration-200 ease-in-out focus:outline-none focus:ring-2 focus:ring-indigo-600 focus:ring-offset-2"
+						role="switch"
+						aria-checked="false"
+						aria-labelledby="darktheme-label"
+						aria-describedby="darktheme-description"
+					>
+						<!-- Enabled: "translate-x-5", Not Enabled: "translate-x-0" -->
+						<span
+							aria-hidden="true"
+							class="translate-x-0 pointer-events-none inline-block h-5 w-5 transform rounded-full bg-white shadow ring-0 transition duration-200 ease-in-out"
+						></span>
+					</button>
+				</div>
+			</div>
+			<div id="tab3" class="tab-content hidden">
+				<ul>
+					<li>
+						<span>Documentation</span>
+						<ul>
+							<li>
+								<a href="https://developers.hubspot.com/docs/cms/hubl/functions"
+									>Functions</a
+								>
+							</li>
+							<li>
+								<a href="https://developers.hubspot.com/docs/cms/hubl/variables"
+									>Variables</a
+								>
+							</li>
+							<li>
+								<a href="https://developers.hubspot.com/docs/cms/hubl/filters"
+									>Filters</a
+								>
+							</li>
+							<li>
+								<a href="https://developers.hubspot.com/docs/cms/data/hubdb"
+									>HubDB</a
+								>
+							</li>
+						</ul>
+					</li>
+					<li>
+						<span>Community</span>
+						<ul>
+							<li>
+								<a href="https://developers.hubspot.com/docs/cms/hubl/functions"
+									>Slack</a
+								>
+							</li>
+							<li><a href="https://community.hubspot.com/">Forum</a></li>
+							<li>
+								<a
+									href="https://community.hubspot.com/t5/HubSpot-Ideas/idb-p/HubSpot_Ideas"
+									>Ideas</a
+								>
+							</li>
+						</ul>
+					</li>
+				</ul>
+			</div>
+		</div>
+
+		<script src="popup.js"></script>
+	</body>
+</html>
